Max 50 scan entries per IP Someone have been adding thousands of players into the queue in matter of seconds. Which is great! We want to have a huge player base! However, this method halts the scanner for ordinary users of the site, having to wait for those thousands of players to finish scanning first.
We already have a backlog of ~59k names if the queue is "empty". If the user who added thousands of names sees this, please reach out and I can add your names into the backlog instead (I have already done this for the ~5k entries you recently did).

If we notice this 50 limit is not respected (constantly filling up 50 entries as soon as possible) so will we probably look into adding a 24 hour IP ban as well.