Scanner Lowered the amount of guaranteed weekly updated users to top 2500 (from 5000).
Simply put, we haven't been able to keep the top 5000 for quite a while. Players entered by visitors on the front page take top priority and new players usually take a lot of scan time to complete.

cPanel (the server software we are using) have made it clear that they won't support "Multiple Network Interfaces" which could solve this issue, so we might be looking into alternatives in the future. If you want to try and change the mind of cPanel, you can always upvote this feature request: https://features.cpanel.net/topic/add-support-for-multiple-network-interfaces