Rework Player ranking is no longer recalculated and stored in the database after each player update. Instead this ranking is calculated "on the fly".
This seems to have made scanning MUCH faster since updating the entire player table every time took some time to execute, while just fetching it "on the fly" is very fast.

On the downside is that game owners and game completions is now only recalculated once an hour - but it's done without locking the database so the scanner can continue to work while this data is being recalculated.